Rooms Added to Phase1 Tester
Copied rooms from Phase1_TSL scene to Phase1_Tester. Converted Ian's door to a prefab in case we need to copy his door in the future. Edited TSL door to deactivate room script, changed tag to door, and renamed the doors in each room to reflect where the doors go to and come from.