Inventory UI Update 1
Added different inventory pannels to switch between melee, ranged, and consumables. inventory also shows ammo for each ranged weapon in the inventory. Slots do not appear if you dont have something in them currently. example if you only had one melee or ranged it would only show the tile for the item you have and the other blank slots do not appear.

Still need to add images for the weapons and consumables in the database.

Still need to add information showcase for each weapon and consumable either on click or hover.