Code for Managing Enemy Spawns
Can now manage enemy waves by adding a EnemySpawnManager object to a room.

EnemySpawnManager stores waves of enemies which include an array for enemies, as well as spawn points

Events are fired that are listed in EventSystem.cs that fire on wave finish, as well as when all waves finished. These can be modified as needed e.g. to include a roomID